Simulation scenario: Three drones moving towards an obstacle(wall) and on the other side of the obstacle, two drones are moving towards each other. One of the 3 drones does not have any detection capability. The second drone can detect anything within a 5 feet distance from it, and when it does it comes to a sudden stop. The third drone can detect anything within 10feet, and when it does- starts moving the other way. The two drones on the other side of the wall should try to avoid the pending collision in someway.
The other 2 drones: Drone 4 has the capability to sense an obstacle or other drone. But it just carries on the same path assuming that the other drone in its path will move. Drone 5 climbs if it senses another drone in its radar, climbs 5feet to avoid the drone and carries on.